requirements.txt
PyAutoGUI~=0.9.54
numpy~=1.25.2
Pillow~=10.0.0